+++ packages/isolate_bloc/lib/src/common/isolate/isolate_manager/web/isolate_manager.dart
@@ -14,8 +14,10 @@ class IsolateManagerImpl extends IsolateManager {
       : super(isolate, messenger);
 
   static Future<IsolateManagerImpl> createIsolate(
-      IsolateRun run, Initializer initializer,
-      [List<String> platformChannels]) async {
+    IsolateRun run,
+    Initializer initializer, [
+    List<String> platformChannels,
+  ]) async {
     final fromIsolate = StreamController<Object>.broadcast();
     final toIsolate = StreamController<Object>.broadcast();
     final sendFromIsolate = fromIsolate.add;